5.49.0 - integrate IADS work by Flogas: Integration of the latest RP fork of the Skynet script. This is identical to the 3.3.0 release, except it also includes this PR.
The 3.3.0 official release includes all the enhancements of the 3.1.5RP that is currently in the VEAF framework (except the PR mentioned above). The code isn't strictly identical, but mostly it is functionally the same.
One notable exception is the HARM defence behavior. In 3.1.5RP it was only disabling radar emissions. In 3.3.0 and 3.3.1RP it will also deactivate the group AI.